수정: 재생 위치 저장이 처음부터 전부 실패하던 문제 + 장 단위 이동 추가
## 재생 위치가 한 번도 저장되지 않고 있었다

라이브러리 진행률이 안 보인다는 제보로 파고들었더니, playback_history
테이블이 통째로 비어 있었다. Fly 로그에 원인이 있었다:

  invalid input syntax for type integer: "78.63075268650299" (22P02)

current_time_seconds 컬럼은 INTEGER인데 클라이언트는 audio.currentTime을
그대로 보낸다. Postgres가 거절하면 서버는 500을 내지만, 클라이언트는 이
저장 실패를 조용히 무시하도록(재생은 계속돼야 하므로) 만들어져 있어
아무도 눈치채지 못했다.

영향 범위가 넓다 — 이어 듣기가 어떤 콘텐츠에서도 동작하지 않았고,
어제 고친 "첫 재생" 지표와 별개로 관리자 대시보드의 재생 관련 수치도
근거 데이터가 없었다. 라이브러리 진행률은 이 위에 얹은 기능이라 함께
동작하지 않았다.

서버에서 초 단위로 자른다. 클라이언트를 고치면 이미 배포된 앱은 계속
실패하므로 서버가 받아주는 쪽이 맞다. 기존 테스트는 전부 정수(120)만
보내서 이 문제를 잡지 못했다 — 실제 프로덕션 값으로 회귀 테스트를 넣었다.

## 라이브러리 진행률이 갱신되지 않던 문제

서점 탭에서 바로 작품을 듣고 리더를 닫으면 탭 전환이 일어나지 않아
카드가 그대로였다. 리더가 닫히는 순간에도 다시 불러온다.

## 장 단위 이동

하단의 ⏮⏭는 10초 이동이라 장 단위 이동 수단이 아예 없었다. 경전처럼
장이 수십 개인 작품에서는 목차가 유일한 통로인데, 현재 위치 표시도
없어 81장 중 어디인지 알 수 없었다.

- 목차에 현재 장 강조 + 열면 그 위치로 자동 스크롤
- 장별 길이 표시(다음 장 시작까지, 마지막 장은 총 재생시간까지)
- 더보기 시트에 이전 장 / 다음 장

"이전 장"은 음악 앱 규칙을 따랐다 — 장을 튼 지 3초 이내면 진짜 이전
장으로, 한참 들었으면 현재 장 처음으로 돌아간다.

백엔드 1개, 프론트 10개 회귀 테스트 추가. 뮤테이션으로 각각 확인했다.

+++ backend/routes/audiobooks.py
@@ -139,7 +139,7 @@ async def update_audiobook(audiobook_id: str, payload: dict, authorization: str 
         raise HTTPException(status_code=500, detail=f"수정에 실패했습니다: {e}")
 
 
-def _validate_playback_state(payload: dict) -> tuple[float, float, str]:
+def _validate_playback_state(payload: dict) -> tuple[int, float, str]:
     position = payload.get("current_time_seconds")
     speed = payload.get("playback_speed", 1.0)
     repeat_mode = payload.get("repeat_mode", "off")
@@ -149,7 +149,13 @@ def _validate_playback_state(payload: dict) -> tuple[float, float, str]:
         raise HTTPException(status_code=400, detail="재생 속도가 올바르지 않습니다.")
     if repeat_mode not in ("off", "all", "one"):
         raise HTTPException(status_code=400, detail="반복 모드가 올바르지 않습니다.")
-    return position, speed, repeat_mode
+    # ⚠️ current_time_seconds 컬럼은 INTEGER다. 클라이언트는 audio.currentTime을
+    # 그대로 보내므로 78.63075268650299 같은 실수가 온다. 이걸 그대로 넘기면
+    # Postgres가 22P02(invalid input syntax for type integer)로 거절하고,
+    # 클라이언트는 저장 실패를 조용히 무시해서 아무도 눈치채지 못한다 —
+    # 그 탓에 playback_history가 통째로 비어 있었다(이어 듣기 불가, 관리자
+    # 대시보드의 재생 지표도 0). 초 단위 정밀도면 충분하므로 여기서 자른다.
+    return int(position), speed, repeat_mode
 
 
 @router.get("/api/audiobooks/{audiobook_id}/playback")
